#43: resolved rebase conflicts

+++ backend/src/main/kotlin/xyz/elitese/ehrenamtskarte/webservice/schema/types/AcceptingStore.kt
@@ -7,24 +7,24 @@ import xyz.elitese.ehrenamtskarte.webservice.dataloader.PHYSICAL_STORE_LOADER_NA
 
 
 data class AcceptingStore(
-        val id: Long,
+        val id: Int,
         val name: String,
-        val contactId: Long,
-        val categoryId: Long
+        val contactId: Int,
+        val categoryId: Int
 ) {
 
     suspend fun physicalStore(dataFetchingEnvironment: DataFetchingEnvironment): PhysicalStore? {
-        return dataFetchingEnvironment.getDataLoader<Long, PhysicalStore?>(PHYSICAL_STORE_LOADER_NAME)
+        return dataFetchingEnvironment.getDataLoader<Int, PhysicalStore?>(PHYSICAL_STORE_LOADER_NAME)
                 .load(id).join()
     }
 
     suspend fun contact(dataFetchingEnvironment: DataFetchingEnvironment): Contact {
-        return dataFetchingEnvironment.getDataLoader<Long, Contact>(CONTACT_LOADER_NAME)
+        return dataFetchingEnvironment.getDataLoader<Int, Contact>(CONTACT_LOADER_NAME)
                 .load(contactId).join()
     }
 
     suspend fun category(dataFetchingEnvironment: DataFetchingEnvironment): Category {
-        return dataFetchingEnvironment.getDataLoader<Long, Category>(CATEGORY_LOADER_NAME)
+        return dataFetchingEnvironment.getDataLoader<Int, Category>(CATEGORY_LOADER_NAME)
                 .load(categoryId).join()
     }
 }
